touchRead() replaced with analogRead()
This allows the library to compile for standard Arduino framework but breaks Capacitive Touch functionality!

We will need to either build in a dependency to an additional library such as "CapacitiveSensor" or allow the user to bring their own capacitive touch library and have the user pass the sensor reading into Flicker.